requirements.txt
numpy>=1.21.0
matplotlib>=3.4.0
pyyaml>=5.4.1
scipy>=1.7.0
tqdm>=4.60.0 